+++ litellm/llms/bedrock/claude_platform/__init__.py
@@ -1,5 +1,7 @@
 from .transformation import (
     BedrockClaudePlatformConfig,
+)
+from .messages_transformation import (
     BedrockClaudePlatformMessagesConfig,
 )
