+++ litellm/llms/bedrock/claude_platform/messages_transformation.py
@@ -1,0 +1,75 @@
+from typing import Any, Dict, List, Optional, Tuple
+
+import litellm
+from litellm.llms.anthropic.experimental_pass_through.messages.transformation import (
+    DEFAULT_ANTHROPIC_API_VERSION,
+    AnthropicMessagesConfig,
+)
+from litellm.secret_managers.main import get_secret_str
+from litellm.types.router import GenericLiteLLMParams
+
+from .common_utils import BedrockClaudePlatformMixin, strip_claude_platform_route
+
+
+class BedrockClaudePlatformMessagesConfig(
+    BedrockClaudePlatformMixin, AnthropicMessagesConfig
+):
+    def validate_anthropic_messages_environment(
+        self,
+        headers: dict,
+        model: str,
+        messages: List[Any],
+        optional_params: dict,
+        litellm_params: dict,
+        api_key: Optional[str] = None,
+        api_base: Optional[str] = None,
+    ) -> Tuple[dict, Optional[str]]:
+        workspace_id = self._get_workspace_id(optional_params, litellm_params)
+        if workspace_id is None:
+            raise litellm.AuthenticationError(
+                message=(
+                    "Missing workspace ID for Claude Platform on AWS. Pass "
+                    "`workspace_id` or configure the provider workspace setting."
+                ),
+                llm_provider="bedrock",
+                model=model,
+            )
+
+        resolved_api_key = (
+            api_key
+            or get_secret_str("ANTHROPIC_AWS_API_KEY")
+            or get_secret_str("ANTHROPIC_API_KEY")
+        )
+        headers = {
+            **headers,
+            "anthropic-version": headers.get(
+                "anthropic-version", DEFAULT_ANTHROPIC_API_VERSION
+            ),
+            "content-type": headers.get("content-type", "application/json"),
+            "anthropic-workspace-id": workspace_id,
+        }
+        if resolved_api_key and "x-api-key" not in headers:
+            headers["x-api-key"] = resolved_api_key
+
+        headers = self._update_headers_with_anthropic_beta(
+            headers=headers,
+            optional_params=optional_params,
+        )
+
+        return headers, api_base
+
+    def transform_anthropic_messages_request(
+        self,
+        model: str,
+        messages: List[Dict],
+        anthropic_messages_optional_request_params: Dict,
+        litellm_params: GenericLiteLLMParams,
+        headers: dict,
+    ) -> Dict:
+        return super().transform_anthropic_messages_request(
+            model=strip_claude_platform_route(model),
+            messages=messages,
+            anthropic_messages_optional_request_params=anthropic_messages_optional_request_params,
+            litellm_params=litellm_params,
+            headers=headers,
+        )
